Fixed app crash bug.
In the basic example, after scanning a code, stopping the scan and then restarting the scan, touching the scan codes table in the view caused the device to crash because the array was empty but the table view hadn’t been refreshed. Added table reloading whenever setting the uniqueCodes array property.