Changed the way we check for scanning availability
Previously we would use simple methods like `scanningIsAvailable` to check for scanning availability. In iOS 8, the app asks permission to use the camera, which doesn't return instantly and may not return successfully.

To account for this, use the `requestCameraPermissionWithSuccess:(void (^)(BOOL success))successBlock` method to check if the app has permission to access the camera.

This will also gracefully handle iOS 7, which doesn't ask for permission before using the camera.